fix(settings): satisfy preset panel lint checks
